requirements.txt
Flask
flask-cors
numpy
scikit-learn
opencv-python
pillow

